Bugzilla Bug #2409: Contribution by Gary Cramblitt and fix confirmed by Charles Stewart
TIFFRGBAImageGet does not properly handle a non-zero col_offset in the
TIFFRGBAImage structure passed as the first argument.  The current (incorrect)
behaviour depends upon the organization of the input tiff file as follows:

1.  If the TIFF is stripped, col_offset is completely ignored.
2.  If the TIFF is tiled, the image result is shifted to the left to the
nearest tile boundary.  For example, if the tilewidth is 100, col_offset is
125, and requested width is 200, you get an image with pixels from 100 up to
300, not 125 up to 325 as expected.

This problem was reported in the mailing list previously as follows, but no
action was taken:

1.  http://www.asmail.be/msg0054834339.html although Andrew is wrong about
tiled images working fine.
2.  http://www.asmail.be/msg0054682793.html